requirements.txt
apipkg==1.5
appdirs==1.4.4
asgiref==3.2.10
atomicwrites==1.4.0
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
coverage==5.1
cryptography==2.9.2
defusedxml==0.6.0
dj-database-url==0.5.0
dj-rest-auth==1.1.1
dj-static==0.0.6
Django==3.1.1
django-allauth==0.42.0
django-cors-headers==3.5.0
django-countries==6.1.3
django-debug-toolbar==2.2
django-environ==0.4.5
django-heroku==0.3.1
django-money==1.1
django-phonenumber-field==5.0.0
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
drf-yasg==1.17.1
execnet==1.7.1
Faker==0.9.1
gunicorn==20.0.4
hypothesis==5.19.2
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mixer==6.1.3
more-itertools==8.4.0
oauthlib==3.1.0
packaging==20.4
pathspec==0.8.0
phonenumberslite==8.12.8
Pillow==7.2.0
pluggy==0.13.1
psycopg2==2.8.5
py==1.9.0
py-moneyed==0.8.0
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.1
pytest-django==3.9.0
pytest-forked==1.2.0
pytest-sugar==0.9.4
pytest-xdist==1.33.0
python-dateutil==2.8.1
python-environ==0.4.54
python3-openid==3.2.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sortedcontainers==2.2.2
sqlparse==0.3.1
static3==0.7.0
termcolor==1.1.0
text-unidecode==1.2
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.5
whitenoise==5.1.0
zipp==3.1.0
